Refactor file structure and update asset paths
