+++ external/jsQR.js
