+++ external/qrcode.js
