+++ offline.html
@@ -142,7 +142,7 @@
   <body>
     <div class="offline-container">
       <img
-        src="/liberdus_logo_250.png"
+        src="./media/liberdus_logo_250.png"
         alt="Liberdus Logo"
         class="offline-logo"
       />
