+++ service-worker.js
@@ -21,8 +21,8 @@ const PRECACHE_URLS = [
   './manifest.json',
   './offline.html',
   './log-utils.js',
-  './liberdus_logo_50.png',
-  './liberdus_logo_250.png'
+  './media/liberdus_logo_50.png',
+  './media/liberdus_logo_250.png'
 ];
 
 // Simplified state management
@@ -498,8 +498,8 @@ async function showNotification(chatCount) {
 
         await self.registration.showNotification('New Messages', {
             body: notificationText,
-            icon: './liberdus_logo_250.png',
-            badge: './liberdus_logo_250.png',
+            icon: './media/liberdus_logo_250.png',
+            badge: './media/liberdus_logo_250.png',
             tag: 'new-messages',
             renotify: true
         });
